+++ src/entity/drawable.rs
@@ -1,4 +1,4 @@
 pub trait Drawable {
     fn draw(&self);
-    fn zdx(&self) -> i32;
+    fn zdx(&self) -> f32;
 }
